+++ src/app/(route)/customer/(route)/feedback/_components/FeedbackItemBox.tsx
@@ -1,8 +1,8 @@
 import { cn } from "@/utils";
 import React from "react";
-import EmptyItem from "../../_components/common/EmptyItem";
-import { FeedbackContentType } from "../../_types/FeedbackItemType";
-import { NoticeContentType } from "../../_types/NoticeItemType";
+import EmptyItem from "../../../_components/common/EmptyItem";
+import { FeedbackContentType } from "../../../_types/FeedbackItemType";
+import { NoticeContentType } from "../../../_types/NoticeItemType";
 import Pagination from "@/app/(route)/mypage/_components/Pagination";
 import { PageInfo } from "@/app/(route)/mypage/_types/toolbarType";
 import { useRouter } from "next/navigation";
