+++ src/app/(route)/customer/(route)/feedback/_components/FeedbackListRender.tsx
@@ -1,7 +1,7 @@
 import FeedbackItem from "./items/FeedbackItem";
 import NoticeItem from "../../notice/_components/items/NoticeItem";
-import { FeedbackContentType } from "../../_types/FeedbackItemType";
-import { NoticeContentType } from "../../_types/NoticeItemType";
+import { FeedbackContentType } from "../../../_types/FeedbackItemType";
+import { NoticeContentType } from "../../../_types/NoticeItemType";
 
 type Props = {
   notices: NoticeContentType[];
