+++ src/app/(route)/customer/(route)/notice/_components/items/NoticeItem.tsx
