+++ src/app/(route)/customer/(route)/notice/_components/status/NoticeItemSkeleton.tsx
