+++ src/app/(route)/customer/(route)/notice/notice-info/[noticeInfoId]/_components/NoticeInfoItemSkeleton.tsx
