+++ src/app/(route)/total-search/_components/SkeletonLoader.tsx
@@ -1,4 +1,4 @@
-import NoticeItemSkeleton from "../../customer/notice/_components/status/NoticeItemSkeleton";
+import NoticeItemSkeleton from "../../customer/(route)/notice/_components/status/NoticeItemSkeleton";
 import NewsPostItemSkeleton from "../../news/_components/NewsPostItemSkeleton";
 
 interface SkeletonLoaderProps {
